Fix double loading of human model texture
The human IQM references the texture models/players/human_male/base
which causes it to be loaded by itself, instead of as part of the
models/players/human_male/body shader as is is normally used. The IQM
materials for player models are not actually used because we always use
skins, but the engine still loads them. On master, this means the
human_male/base texture is loaded twice. (In 0.55.4 it is only loaded
once, but the imageMinDimension in the shader is ignored. You can force
the old behavior by setting r_allowImageParamMismatch.)

Fix this by setting the IQM material to the shader
models/players/human_male/body, so that we don't load a variant of the
material that is never used.